% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check function input data}
\usage{
check_data(
  n = NA,
  SD = NA,
  SDp = NA,
  SEp = NA,
  SE = NA,
  n1 = NA,
  n2 = NA,
  SD1 = NA,
  SD2 = NA,
  SD_diff = NA,
  r = NA,
  SD_within = NA,
  CI_low = NA,
  CI_up = NA,
  OR = NA,
  sig_level = NA,
  p = NA
)
}
\description{
Check function input data
}
\keyword{internal}
