% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyforest.R
\name{easyforest}
\alias{easyforest}
\title{Create forest plots for metafor outcomes}
\usage{
easyforest(
  rma_results,
  sig_level = 0.05,
  null_value = 0,
  margin_bottom = 1,
  headlines = "",
  error_bar_height = 0.2,
  pooled_estimate_height = 0.3,
  textsize = 3,
  title_size = 3.5,
  top_space = 0,
  pooled_estimate_position = -1,
  distance_headline = 1.4,
  group_distance = 1.5,
  show_fav = FALSE,
  fav_height = 0.017,
  fav_space = 0.05,
  fav_textsize = 2.5,
  fav_lab = c("Favours X", "Favours Y"),
  CI_title = "SMD [95\% CI]",
  label_title = "Author (Year)",
  values_title = "Standardized Mean Differences",
  summary_title = "Random Effects"
)
}
\arguments{
\item{rma_results}{a list of rma results created by metafor}

\item{sig_level}{the two sided significance value}

\item{null_value}{null effect size. In case of SMD = 0 and RR = 1}

\item{margin_bottom}{distance from last effect description to x_axis}

\item{headlines}{group headlines}

\item{error_bar_height}{height of the errorbars}

\item{pooled_estimate_height}{height of the pooled estimate parallelogram}

\item{textsize}{general text size}

\item{title_size}{size of the plot title}

\item{top_space}{space from panel to top}

\item{pooled_estimate_position}{position of the pooled effect in relation to the lowermost study}

\item{distance_headline}{distance of the group headlines}

\item{group_distance}{distance from one group to the next}

\item{CI_title}{title for the text CIs}

\item{label_title}{title for the studies}

\item{values_title}{title for the plotted CIs}

\item{summary_title}{title written in front of the summary effects}

\item{z}{transforms the sig_level to z}
}
\value{
Forest Plot
}
\description{
Create forest plots for metafor outcomes
}
\examples{
### copy BCG vaccine meta-analysis data into 'dat'

library(metafor)
dat <- dat.bcg

### calculate log risk ratios and corresponding sampling variances (and use
### the 'slab' argument to store study labels as part of the data frame)
dat <- metafor::escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat,
             slab=paste(author, year, sep=", "))

### fit random-effects model
res <- metafor::rma(yi, vi, data=dat)

#' easyforest(res)
}
\keyword{internal}
