% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_var.R
\name{extend_var}
\alias{extend_var}
\title{Extends the length of a variable}
\usage{
extend_var(variable, length)
}
\arguments{
\item{variable}{variable that should be extended}

\item{length}{length of the returned variable}
}
\value{

}
\description{
Extends the length of a variable
}
\examples{
x <- "test"
extend_var(x, 5)
}
\keyword{internal}
