% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_factor.R
\name{hedges_factor}
\alias{hedges_factor}
\title{Hedges factor calculation}
\usage{
hedges_factor(n1 = NA, n2 = NA, n_total = NA)
}
\arguments{
\item{n1}{sample size group 1}

\item{n2}{sample size group 2}

\item{n_total}{total sample size}
}
\value{

}
\description{
Hedges factor calculation
}
\references{
Hedges L. V., Olkin I. (1985). Statistical methods for meta-analysis. San Diego, CA: Academic Press
}
