% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_factor.R
\name{hedges_factor_approx}
\alias{hedges_factor_approx}
\title{Computational simple calculation of the hedges factor}
\usage{
hedges_factor_approx(n1 = NA, n2 = NA, n_total = NA)
}
\arguments{
\item{n1}{sample size group 1}

\item{n2}{sample size group 2}

\item{n_total}{total sample size}
}
\value{

}
\description{
Computational simple calculation of the hedges factor
}
\references{
Borenstein, M., Hedges, L.V., Higgins, J.P.T. and Rothstein, H.R. (2009). Effect Sizes Based on Means. In Introduction to Meta-Analysis (eds M. Borenstein, L.V. Hedges, J.P.T. Higgins and H.R. Rothstein). https://doi.org/10.1002/9780470743386.ch4
}
