% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{r.pre_post}
\alias{r.pre_post}
\title{Approximates correlation coefficient pre post studies}
\usage{
r.pre_post(SD_diff, r)
}
\arguments{
\item{SD_diff}{standard deviation of the difference}

\item{r}{correlation between pair of observations}
}
\value{
Within standard deviation
}
\description{
The correlation correlation coefficient can then be used to calculate SD for
mean changes.
}
\examples{
# SD_diff is the standard deviation of the group difference
SD_diff <- 2
# r is the correlation coefficient between the groups
r <- 0.5
SD.within_from_SD.r(SD_diff, r)
}
\references{

}
