% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SD.within_from_SD.r}
\alias{SD.within_from_SD.r}
\title{'Within Standard Deviation for Matched Groups}
\usage{
SD.within_from_SD.r(SD_diff, r)
}
\arguments{
\item{SD_diff}{standard deviation of the difference (usually reported)}

\item{r}{correlation between pair of observations}
}
\value{
Within standard deviation
}
\description{
Computes the within standard deviation for matched groups. The within standard deviation can further be used to compute the standardized mean
differences for matched groups.
}
\examples{
# SD_diff is the standard deviation of the group difference
SD_diff <- 2
# r is the correlation coefficient between the groups
r <- 0.5
SD.within_from_SD.r(SD_diff, r)
}
