% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SD_M_n_pooled_from_groups}
\alias{SD_M_n_pooled_from_groups}
\title{Combined Standard Deviation for Multiple Groups}
\usage{
SD_M_n_pooled_from_groups(M, SD, n)
}
\arguments{
\item{M}{vector of group means}

\item{SD}{vector of group SDs}

\item{n}{vector of group sample sizes}
}
\value{
Within standard deviation
}
\description{
Computes the pooled standard deviation for multiple groups.
}
\details{
This function also returns the combined mean and the total sample size across all groups.
Use this function to combine subgroups into one combined group. This requires subgroup means, SDs, and sample sizes. \code{\link[=SDp_from_SD]{SDp_from_SD()}} calculates a pooled within group SD denominator for two independent groups and is not the same as combining subgroups into one group when subgroup means differ.
}
\examples{
# Compute the Standard deviation for the following grouped data
M <- c(1, 1.5, 2) # Means
SD <- c(2, 3, 2.5) # SDs
n <- c(72, 80, 55) # sample sizes
SD_M_n_pooled_from_groups(M, SD, n)
}
\references{
\href{https://www.cochrane.org/authors/handbooks-and-manuals/handbook/archive/v5.1.0}{Cochrane Handbook}

Rücker G, Cates CJ, Schwarzer G. Methods for including information from multi-arm trials in pairwise meta-analysis. Res Synth Methods. 2017 Dec;8(4):392-403. doi: 10.1002/jrsm.1259. Epub 2017 Aug 25. PMID: 28759708.
}
