% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SD_pool_hedge}
\alias{SD_pool_hedge}
\title{Pools the SD according to hedge}
\usage{
SD_pool_hedge(SD1, SD2, n1, n2)
}
\arguments{
\item{SD2}{}
}
\value{

}
\description{
Pools the SD according to hedge
}
\keyword{internal}
