% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SD_pool_hedges}
\alias{SD_pool_hedges}
\title{Pools the SD according to hedge}
\usage{
SD_pool_hedges(SD1, SD2, n1, n2)
}
\arguments{
\item{SD1}{standard deviation group 1}

\item{SD2}{standard deviation group 2}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\value{

}
\description{
Pools the SD according to hedge
}
