% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SDp_from_SEp}
\alias{SDp_from_SEp}
\title{Pooled Within-Group Standard Deviation from SE of Raw Mean Difference}
\usage{
SDp_from_SEp(SEp, n1, n2)
}
\arguments{
\item{SEp}{SE of a raw mean difference from two independent groups}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\value{
Pooled standard deviation
}
\description{
\strong{IMPORTANT}: For a single group, use \code{\link[=SD_from_SE]{SD_from_SE()}}!
Calculates the implied pooled within-group standard deviation from the SE of a raw mean difference from two independent groups and their sample sizes. This method is the reverse of \code{\link[=SEp_from_SDp]{SEp_from_SDp()}} under the equal SD assumption.
Do not use with SEs for SMDs, odds ratios, risk ratios, hazard ratios, log effects, adjusted model estimates when reconstructing raw SDs, Welch tests, or paired designs.
}
\examples{
# SE of a raw mean difference, sample size 1 and sample size 2
SE <- 0.12
n1 <- 140
n2 <- 140

SDp_from_SEp(SE, n1, n2)
}
\references{
\href{https://handbook-5-1.cochrane.org/chapter_7/7_7_3_3_obtaining_standard_deviations_from_standard_errors.htm}{Cochrane Handbook}
}
\seealso{
\code{\link[=SD_from_SE]{SD_from_SE()}} for a single group.
\code{\link[=SEp_from_SDp]{SEp_from_SDp()}} if the standard error should be computed instead.
}
