% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE.SMD_from_SMD}
\alias{SE.SMD_from_SMD}
\title{SMD standard error from sample sizes}
\usage{
SE.SMD_from_SMD(SMD, n1, n2, method = "hedges")
}
\arguments{
\item{SMD}{standardized mean differences}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\value{
Standard error of SMD (e.g. standard error of intervention effect)
}
\description{
Approximates SMD standard error from sample sizes. Approximation provided by Borenstein (2009, p.226).
}
\examples{
# SMD = 0.6, sample size group_1 = 50, sample size group_2 = 75
SE.SMD_from_SMD(0.6, 50, 75)
}
\references{
Borenstein M: Effect sizes for continuous data.
The Handbook of Research Synthesis and Meta-Analysis. Edited by: Cooper H, Hedges LV, Valentine JC. 2009,
New York: Russell Sage, 221-235. 2
}
