% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE.SMD_from_SMD_n}
\alias{SE.SMD_from_SMD_n}
\title{Approximate SMD standard error from sample sizes}
\usage{
SE.SMD_from_SMD_n(n1, n2, SMD)
}
\arguments{
\item{n1}{sample size group 1}

\item{n2}{sample size group 2}

\item{SMD}{standardized mean differences}
}
\value{
Returns standard error
}
\description{
Estimate SMD standard error from sample sizes. Approximation provided by Borenstein (2009, p.226)
}
\details{
Literature:
Borenstein M: Effect sizes for continuous data.
The Handbook of Research Synthesis and Meta-Analysis. Edited by: Cooper H, Hedges LV, Valentine JC. 2009,
New York: Russell Sage, 221-235. 2
}
