% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE_from_CI}
\alias{SE_from_CI}
\title{Obtains standard error from confidence intervals}
\usage{
SE_from_CI(CI_low, CI_up, sig_level = 0.05, two_sided = TRUE)
}
\arguments{
\item{CI_low}{lover OR confidence interval limit}

\item{CI_up}{upper OR confidence interval limit}

\item{sig_level}{the significance level}

\item{two_sided}{whether the two sided z statistics or single sided should be calculated}
}
\value{

}
\description{
Method is according to the archived version of the Cochrane handbook:
https://handbook-5-1.cochrane.org/chapter_7/7_7_7_2_obtaining_standard_errors_from_confidence_intervals_and.htm
}
