% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE_from_SD.N}
\alias{SE_from_SD.N}
\title{Calculates the Standard Error for a Single Group}
\usage{
SE_from_SD.N(SD, n)
}
\arguments{
\item{SD}{standard deviation}

\item{n}{sample size

Calculates the standard error with the following formulae:
SD / sqrt(n)}
}
\value{

}
\description{
Calculates the Standard Error for a Single Group
}
