% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SEp_from_SDp.n}
\alias{SEp_from_SDp.n}
\title{Pooled Standard Error}
\usage{
SEp_from_SDp.n(SDp, n1, n2)
}
\arguments{
\item{SDp}{pooled standard deviation}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\value{
Pooled standard error for for two groups
}
\description{
Caclulates the pooled standard error for two groups (e.g. intervention effect). In case of a single group, \code{\link[=SE_from_SD.N]{SE_from_SD.N()}} has to be used.
}
\examples{
# Pooled standard deviation = 2, sample size group a = 50, sample size group b = 75,
SEp_from_SDp.n(2, 50, 75)
}
\references{
https://handbook-5-1.cochrane.org/chapter_7/7_7_3_3_obtaining_standard_deviations_from_standard_errors.htm
}
\seealso{
\code{\link[=SEp_from_SDp.N]{SEp_from_SDp.N()}} for a single group
}
