% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SEp_from_SDp}
\alias{SEp_from_SDp}
\title{Standard Error (Pooled)}
\usage{
SEp_from_SDp(SDp, n1, n2)
}
\arguments{
\item{SDp}{pooled standard deviation}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\value{
Pooled standard error for two groups (e.g. standard error of intervention effect)
}
\description{
\strong{IMPORTANT}: When there is only one group, the following method has to be used: \code{\link[=SE_from_SD]{SE_from_SD()}}
Calculates the pooled standard error for two groups (e.g., intervention effect).
}
\examples{
# Pooled standard deviation = 2, sample size group a = 50, sample size group b = 75
SEp_from_SDp(2, 50, 75)
}
\references{
\href{https://www.cochrane.org/authors/handbooks-and-manuals/handbook/archive/v5.1.0}{Cochrane Handbook}
}
\seealso{
\code{\link[=SE_from_SD]{SE_from_SD()}} for a single group
}
