% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{SMD.matched_calc}
\alias{SMD.matched_calc}
\title{Calculates SMD from matched groups}
\usage{
SMD.matched_calc(M_diff = NA, M1 = NA, M2 = NA, SD_within)
}
\arguments{
\item{M_diff}{mean difference between groups}

\item{M1}{mean group 1 (in case M_diff not provided)}

\item{M2}{mean group 2 (in case M_diff not provided)}

\item{SD_within}{within standard deviation. CAVE this is usually not reported but needs to be computed from the difference standard deviation.
This can be done with SD.within_from_SD.r().}
}
\value{

}
\description{
Calculates the standardized mean differences for matched groups. Needs either the mean of the groups or
the difference between groups.
SD_within is usually not reported but can be calculated by the use of SD.within_from_SD.r()
}
\examples{
# Calcuation with group means
SMD.matched_calc(M1 = 103, M2 = 100, SD_within = 7.1005)

# Calculation with group difference
SMD.matched_calc(M_diff = 3, SD_within = 7.1005)

# Calculation with standard deviation between
# Correlation Coefficient between groups
r <- 0.7

# SD between groups
SD_between <- 5.5

SMD.matched_calc(M_diff = 3,
    SD_within = SD.within_from_SD.r(SD_between, r))
}
\references{
M., Hedges, L.V., Higgins, J.P.T. and Rothstein, H.R. (2009). Converting Among Effect Sizes. In Introduction to Meta-Analysis (eds M. Borenstein, L.V. Hedges, J.P.T. Higgins and H.R. Rothstein). https://doi.org/10.1002/9780470743386.ch7
}
