% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{SMD.SE_from_OR.CI}
\alias{SMD.SE_from_OR.CI}
\title{SE of SMD from Confidence Intervals of OR}
\usage{
SMD.SE_from_OR.CI(CI_low, CI_up, z = 1.96)
}
\arguments{
\item{CI_low}{lover OR confidence interval limit}

\item{CI_up}{upper OR confidence interval limit}

\item{z}{z-value depending on the significance level (0.05 two-sided = 1.96)}
}
\value{

}
\description{
SE of SMD from Confidence Intervals of OR
}
