% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{SMD_calc}
\alias{SMD_calc}
\title{Calculates SMD}
\usage{
SMD_calc(M1, M2, SD_pooled)
}
\arguments{
\item{M1}{treatment effect size group 1}

\item{M2}{treatment effect size group 2}

\item{SD_pooled}{the pooled standard deviation or the standard deviation of the control group in case Glass's delta should be calculated}
}
\value{

}
\description{
Uses the differences of two means and divides with the (pooled) standard deviation or the standard deviation of the
control group in case Glass's delta should be calculated.
}
