% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{SMD_from_arm}
\alias{SMD_from_arm}
\title{Calculates SMD from arm data}
\usage{
SMD_from_arm(M1, M2, SD1, SD2, n1 = NA, n2 = NA, method = "hedges")
}
\arguments{
\item{M1}{treatment effect size group 1}

\item{M2}{treatment effect size group 2}

\item{SD1}{standard deviation group 1}

\item{SD2}{standard deviation group 2}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}

\item{method}{}
}
\value{

}
\description{
Calculates SMD from arm data
}
