% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_calc.R
\name{t_calc}
\alias{t_calc}
\title{Calculates the t_value}
\usage{
t_calc(sig_level = 0.05, two_sided = TRUE, df)
}
\arguments{
\item{sig_level}{the significance level}

\item{two_sided}{whether the two sided t statistics or single sided should be calculated}

\item{df}{= degrees of freedom}
}
\description{
Calculates the t_value
}
