% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_calc.R
\name{z_calc}
\alias{z_calc}
\title{Calculates the z_value}
\usage{
z_calc(sig_level = 0.05, two_sided = TRUE)
}
\arguments{
\item{sig_level}{the significance level}

\item{two_sided}{whether the two sided z statistics or single sided should be calculated}
}
\value{

}
\description{
Calculates the z_value
}
