% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SD_from_CI}
\alias{SD_from_CI}
\title{Standard Deviation from Confidence interval}
\usage{
SD_from_CI(CI_low, CI_up, N, sig_level = 0.05, two_sided = TRUE)
}
\value{
Standard deviation single group
}
\description{
Computes the standard deviation from the confidence interval and sample size. This method is only valid for single groups.
For two groups (e.g. intervention effects) SDp_from_CIp has to be used.
}
\references{
https://handbook-5-1.cochrane.org/chapter_7/7_7_3_2_obtaining_standard_deviations_from_standard_errors_and.htm
}
