% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SD_from_SE}
\alias{SD_from_SE}
\title{SD from Standard Error}
\usage{
SD_from_SE(SE, n)
}
\arguments{
\item{SE}{standard error}

\item{n}{sample size}
}
\value{

}
\description{
This is a simple helper to calculate standard deviation from the standard error
}
