% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SD_pool}
\alias{SD_pool}
\title{Pooled SD from two SDs}
\usage{
SD_pool(SD1, SD2, n1 = NA, n2 = NA, method = "hedges")
}
\arguments{
\item{SD1}{Standard deviation of group 1}

\item{SD2}{Standard deviation of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}

\item{method}{the method that should be used to calculate the SD. "hedge" calculates SD* Hedges' (1981, p.110).
"cohen" uses the simplified method by Cohen 1988,}
}
\value{

}
\description{
Calculates pooled standard deviation. The method according to Hedges (1981, p.110)
needs the sample sizes (https://www.polyu.edu.hk/mm/effectsizefaqs/effect_size_equations2.html).
If exclusively SDs are provided, the simpler equation provided by Cohen 1988 can be used.
}
\references{
https://www.polyu.edu.hk/mm/effectsizefaqs/effect_size_equations2.html

Hedges, L. V. (1981). Distribution theory for Glass’s estimator of effect size and related estimators.
Journal of Educational Statistics, 6, 107–128.

https://www.meta-analysis.com/downloads/Meta-analysis\%20Effect\%20sizes\%20based\%20on\%20means.pdf
}
