% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SDp_from_SEp}
\alias{SDp_from_SEp}
\title{Calculates the pooled standard deviation from the pooled standard error}
\usage{
SDp_from_SEp(SEp, n1, n2)
}
\arguments{
\item{n2}{}
}
\value{

}
\description{
It is the reverse method of SEp_from_SDp.N()
}
\details{
Literature:
https://handbook-5-1.cochrane.org/chapter_7/7_7_3_3_obtaining_standard_deviations_from_standard_errors.htm
}
