% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD.R
\name{SDp_from_SEp}
\alias{SDp_from_SEp}
\title{Standard Deviation from the Pooled Standard Error}
\usage{
SDp_from_SEp(SEp, n1, n2)
}
\arguments{
\item{SEp}{pooled standard error}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\value{
Pooled standard deviation
}
\description{
In case studies provide the standard error of a treatment effect estimate of two groups.
It is the reverse method of SEp_from_SDp.N()
}
\examples{
#pooled standard error, sample size 1 and samplie size 2
SE <- 0.12
n1 <- 140
n2 <- 140

SDp_from_SEp(SE, n1, n2)
}
\references{
https://handbook-5-1.cochrane.org/chapter_7/7_7_3_3_obtaining_standard_deviations_from_standard_errors.htm
}
