% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE.SMD_from_OR.CI}
\alias{SE.SMD_from_OR.CI}
\title{SE of SMD from Confidence Intervals of Odds Ratio (OR)}
\usage{
SE.SMD_from_OR.CI(CI_low, CI_up, sig_level = 0.05, two_sided = TRUE)
}
\arguments{
\item{CI_low}{lower odds ratio confidence interval limit}

\item{CI_up}{upper odds ratio confidence interval limit}

\item{sig_level}{the significance level}

\item{two_sided}{whether the two sided z statistics or single sided should be calculated}
}
\value{

}
\description{
https://stats.stackexchange.com/questions/68290/converting-odds-ratios-to-cohens-d-for-meta-analysis
}
\details{
The method uses multiple steps in background:
\enumerate{
\item Takes OR limits and transforms them to log(OR)
\item Calculates the standard error for the log(OR)
\item Transforms the log(OR) standard error to SMD standard error by multiplying it with sqrt(3)/pi
}
}
\examples{
# lower CI = 0.6, upper CI = 0.9
SE.SMD_from_OR.CI(0.6, 0.9)
}
\references{
Chinn S. A simple method for converting an odds ratio to effect size for use in meta-analysis. Stat Med. 2000 Nov 30;19(22):3127-31. doi: 10.1002/1097-0258(20001130)19:22<3127::aid-sim784>3.0.co;2-m. PMID: 11113947.
}
