% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE_from_SD}
\alias{SE_from_SD}
\title{Standard Error for a Single Group}
\usage{
SE_from_SD(SD, n)
}
\arguments{
\item{SD}{standard deviation}

\item{n}{sample size}
}
\value{
Single group standard error
}
\description{
Calculates the standard error for a single group. In case of two groups (e.g. intervention effect), \code{\link[=SEp_from_SDp]{SEp_from_SDp()}} has to be used.
}
\examples{
# Standard deviation = 2, group size = 50
SE_from_SD(2, 50)
}
\references{
https://handbook-5-1.cochrane.org/chapter_7/7_7_3_2_obtaining_standard_deviations_from_standard_errors_and.htm#'
}
\seealso{
\code{\link[=SEp_from_SDp]{SEp_from_SDp()}} for two groups
}
