% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SEp_from_CIp}
\alias{SEp_from_CIp}
\title{Standard Error from Confidence Interval for differences of means}
\usage{
SEp_from_CIp(
  CI_low,
  CI_up,
  N1 = NA,
  N2 = NA,
  sig_level = 0.05,
  two_sided = TRUE,
  t_dist = TRUE
)
}
\arguments{
\item{CI_low}{lover OR confidence interval limit}

\item{CI_up}{upper OR confidence interval limit}

\item{sig_level}{the significance level}

\item{two_sided}{whether the two sided statistics or single sided should be calculated}
}
\value{
Pooled standard error (e.g. intervention effect)
}
\description{
Calculates the standard error from confidence interval limits for differences of means.
This method is only valid in case the confidence interval is symmetrical
around the mean and works only for the t or normal distribution (argument t_dist).
}
\examples{
# lower CI = -1.5, upper CI = 0.5
SEp_from_CIp(-1.5, 0.5)
}
\references{
https://handbook-5-1.cochrane.org/chapter_7/7_7_7_2_obtaining_standard_errors_from_confidence_intervals_and.htm
}
