% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SEp_from_SDp.N}
\alias{SEp_from_SDp.N}
\title{Calculates the pooled Standard Error}
\usage{
SEp_from_SDp.N(SDp, n1, n2)
}
\arguments{
\item{SDp}{pooled standard deviation}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}
}
\value{

}
\description{
If SMD are used it is recommended to use the approximation provided by Borenstein:
See ?SE.SMD_from_SMD.n
}
\references{
https://handbook-5-1.cochrane.org/chapter_7/7_7_3_3_obtaining_standard_deviations_from_standard_errors.htm
}
