% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SEp_from_TE.p}
\alias{SEp_from_TE.p}
\title{Standard Error from Treatment Effect and p-value}
\usage{
SEp_from_TE.p(TE, p, two_sided = TRUE)
}
\arguments{
\item{TE}{reported treatment effect}

\item{p}{reported p-value}

\item{two_sided}{whether the two sided z statistics or single sided should be calculated}
}
\value{
Pooled standard error
}
\description{
Calculates the pooled standard error by the use of the treatment effect and p value.
To avoid an infinitive return in case of p = 1, p is automatically adjusted to 0.99999
}
\examples{
# TE = 1.5, p = 0.8
SEp_from_TE.p(1.5, 0.8)
}
\references{
Altman D G, Bland J M. How to obtain the confidence interval from a P value BMJ 2011; 343 :d2090 doi:10.1136/bmj.d2090
}
