% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{SMD_from_OR}
\alias{SMD_from_OR}
\title{Convert SMD to OR}
\usage{
SMD_from_OR(OR)
}
\arguments{
\item{OR}{Odds Ratio}
}
\value{
Returns SMD
}
\description{
Converts OR to SMD by using the formula provided here:
https://stats.stackexchange.com/questions/68290/converting-odds-ratios-to-cohens-d-for-meta-analysis
}
\details{
Borenstein, M., Hedges, L.V., Higgins, J.P.T. and Rothstein, H.R. (2009). Converting Among Effect Sizes. In Introduction to Meta-Analysis (eds M. Borenstein, L.V. Hedges, J.P.T. Higgins and H.R. Rothstein). https://doi.org/10.1002/9780470743386.ch7
}
\examples{

}
